Break out effects, mail, npcs, and the rest of the functions in Format
