+++ test/game/format/skills_test.exs
@@ -1,0 +1,7 @@
+defmodule Game.Format.SkillsTest do
+  use ExUnit.Case
+
+  alias Game.Format.Skills
+
+  doctest Game.Format.Skills
+end
