+++ test/game/format_test.exs
@@ -32,45 +32,4 @@ defmodule Game.FormatTest do
       assert Format.wrap("hi\n\n\nthere") == "hi\n\n\nthere"
     end
   end
-
-  describe "shop listing" do
-    setup do
-      sword = %{name: "Sword", price: 100, quantity: 10}
-      shield = %{name: "Shield", price: 80, quantity: -1}
-      %{shop: %{name: "Tree Top Stand"}, items: [sword, shield]}
-    end
-
-    test "includes shop name", %{shop: shop, items: items} do
-      assert Regex.match?(~r/Tree Top Stand/, Format.list_shop(shop, items))
-    end
-
-    test "includes shop items", %{shop: shop, items: items} do
-      assert Regex.match?(~r/100 gold/, Format.list_shop(shop, items))
-      assert Regex.match?(~r/10 left/, Format.list_shop(shop, items))
-      assert Regex.match?(~r/Sword/, Format.list_shop(shop, items))
-    end
-
-    test "-1 quantity is unlimited", %{shop: shop, items: items} do
-      assert Regex.match?(~r/unlimited/, Format.list_shop(shop, items))
-    end
-  end
-
-  describe "npc status line" do
-    setup do
-      npc = %{name: "Guard", extra: %{status_line: "[name] is here.", is_quest_giver: false}}
-
-      %{npc: npc}
-    end
-
-    test "templates the name in", %{npc: npc} do
-      assert Format.npc_name_for_status(npc) == "{npc}Guard{/npc}"
-      assert Format.npc_status(npc) == "{npc}Guard{/npc} is here."
-    end
-
-    test "if a quest giver it includes a quest mark", %{npc: npc} do
-      npc = %{npc | extra: Map.put(npc.extra, :is_quest_giver, true)}
-      assert Format.npc_name_for_status(npc) == "{npc}Guard{/npc} ({quest}!{/quest})"
-      assert Format.npc_status(npc) == "{npc}Guard{/npc} ({quest}!{/quest}) is here."
-    end
-  end
 end
